Refactor storage layer and update service protocols
- Removed the old storage package and replaced it with a new store package.
- Updated service protocol types from storage.ServiceProtocolType to models.ServiceProtocolType.
- Changed the retries field type from int to int64 in various structs and database models.
- Refactored the TestSuite and MonitorService to use the new store package.
- Implemented incident statistics retrieval by date range in the incidents service.
- Updated web handlers and responses to reflect changes in the storage and service models.
- Removed unused storage error handling and incident management code.

+++ internal/models/service.go
@@ -33,7 +33,7 @@ type ServiceFullView struct {
 	Protocol           ServiceProtocolType `json:"protocol"`
 	Interval           time.Duration       `json:"interval" swaggertype:"primitive,integer"`
 	Timeout            time.Duration       `json:"timeout" swaggertype:"primitive,integer"`
-	Retries            int                 `json:"retries"`
+	Retries            int64               `json:"retries"`
 	Tags               []string            `json:"tags"`
 	Config             map[string]any      `json:"config"`
 	IsEnabled          bool                `json:"is_enabled"`
